feat: add new obscured document lenses (Lattice, Chronos, Ethereal)
- Registers three new lenses using `InputManager` in `src/interactions/lensBindings.js`:
  - Prismatic Lattice (`Alt+Shift+[`)
  - Chronos Pulse (`Alt+Shift+]`)
  - Ethereal Cascade (`Alt+Shift+'`)
- Appends styling rules (`.prismatic-lattice-active`, `.chronos-pulse-active`, `.ethereal-cascade-active`) to `src/styles_ui.css` applying custom 3D transforms, masks, and CSS filters dynamically via `--item-index`.
- Verified layout logic correctly preserves other view mode interactions.
