+++ src/styles_ui.css
@@ -2452,3 +2452,128 @@ body.nebula-core-active .echo-document::before {
 @keyframes nebula-core-spin {
   100% { transform: rotate(360deg); }
 }
+
+/* ─── Prismatic Lattice Lens (Alt+Shift+[) ────────────────────────────── */
+body.prismatic-lattice-active #editor {
+  mask-image: repeating-linear-gradient(
+    45deg,
+    transparent,
+    transparent 10px,
+    rgba(0,0,0,0.8) 10px,
+    rgba(0,0,0,0.8) 20px
+  ), repeating-linear-gradient(
+    -45deg,
+    transparent,
+    transparent 10px,
+    rgba(0,0,0,0.8) 10px,
+    rgba(0,0,0,0.8) 20px
+  );
+  mask-composite: intersect;
+  -webkit-mask-image: repeating-linear-gradient(
+    45deg,
+    transparent,
+    transparent 10px,
+    rgba(0,0,0,0.8) 10px,
+    rgba(0,0,0,0.8) 20px
+  ), repeating-linear-gradient(
+    -45deg,
+    transparent,
+    transparent 10px,
+    rgba(0,0,0,0.8) 10px,
+    rgba(0,0,0,0.8) 20px
+  );
+  -webkit-mask-composite: source-in;
+  opacity: 0.7;
+  transition: opacity 0.3s ease;
+}
+
+body.prismatic-lattice-active .echo-document {
+  transition: transform 0.5s cubic-bezier(0.2, 0.8, 0.2, 1), filter 0.5s ease;
+  transform: translate3d(
+    calc(var(--tx, 0px) + (var(--item-index, 0) - 2) * 40px),
+    calc(var(--ty, 0px) + (var(--item-index, 0) - 2) * -30px),
+    calc(var(--tz, 0px) + 150px - (var(--item-index, 0) * 50px))
+  ) rotateX(calc(var(--item-index, 0) * 15deg)) rotateY(calc(var(--item-index, 0) * -15deg)) scale(calc(1.1 - (var(--item-index, 0) * 0.05))) !important;
+  filter: hue-rotate(calc(var(--item-index, 0) * 60deg)) saturate(200%) brightness(1.2) drop-shadow(0 0 15px rgba(255, 0, 255, 0.5)) !important;
+  z-index: calc(200 - var(--item-index, 0)) !important;
+  opacity: 0.85 !important;
+  border: 1px solid rgba(255, 0, 255, 0.4) !important;
+  box-shadow: inset 0 0 20px rgba(255, 0, 255, 0.2) !important;
+}
+
+
+/* ─── Chronos Pulse Lens (Alt+Shift+]) ────────────────────────────────── */
+body.chronos-pulse-active #editor {
+  mask-image: radial-gradient(
+    circle 200px at var(--lens-x, 50%) var(--lens-y, 50%),
+    transparent 0%,
+    rgba(0,0,0,0.4) 30%,
+    black 80%
+  );
+  -webkit-mask-image: radial-gradient(
+    circle 200px at var(--lens-x, 50%) var(--lens-y, 50%),
+    transparent 0%,
+    rgba(0,0,0,0.4) 30%,
+    black 80%
+  );
+  animation: chronos-pulse-mask 2s infinite alternate;
+}
+
+@keyframes chronos-pulse-mask {
+  0% { opacity: 0.8; }
+  100% { opacity: 0.4; }
+}
+
+body.chronos-pulse-active .echo-document {
+  transition: transform 0.6s cubic-bezier(0.4, 0, 0.2, 1), filter 0.6s ease;
+  transform: translate3d(
+    var(--tx, 0px),
+    var(--ty, 0px),
+    calc(var(--tz, 0px) - (var(--item-index, 0) * 80px))
+  ) scale(calc(1 + (var(--item-index, 0) * 0.1))) !important;
+  filter: sepia(0.8) hue-rotate(-30deg) contrast(1.5) blur(calc(var(--item-index, 0) * 1.5px)) drop-shadow(0 0 20px rgba(255, 50, 50, 0.6)) !important;
+  z-index: calc(100 - var(--item-index, 0)) !important;
+  opacity: 0.8 !important;
+  border: 2px solid rgba(255, 50, 50, 0.3) !important;
+  animation: chronos-ripple 3s infinite;
+  animation-delay: calc(var(--item-index, 0) * 0.2s);
+}
+
+@keyframes chronos-ripple {
+  0% { box-shadow: 0 0 0 0 rgba(255, 50, 50, 0.4); }
+  70% { box-shadow: 0 0 0 30px rgba(255, 50, 50, 0); }
+  100% { box-shadow: 0 0 0 0 rgba(255, 50, 50, 0); }
+}
+
+
+/* ─── Ethereal Cascade Lens (Alt+Shift+') ─────────────────────────────── */
+body.ethereal-cascade-active #editor {
+  mask-image: linear-gradient(
+    to bottom,
+    transparent 0%,
+    rgba(0,0,0,0.5) 40%,
+    black 100%
+  );
+  -webkit-mask-image: linear-gradient(
+    to bottom,
+    transparent 0%,
+    rgba(0,0,0,0.5) 40%,
+    black 100%
+  );
+  opacity: 0.6;
+  transition: opacity 0.4s ease;
+}
+
+body.ethereal-cascade-active .echo-document {
+  transition: transform 0.8s cubic-bezier(0.25, 0.1, 0.25, 1), filter 0.8s ease;
+  transform: translate3d(
+    calc(var(--tx, 0px) + (sin(calc(var(--item-index, 0) * 45deg)) * 30px)),
+    calc(var(--ty, 0px) + (var(--item-index, 0) * 50px)),
+    calc(var(--tz, 0px) - (var(--item-index, 0) * 60px))
+  ) rotateZ(calc(var(--item-index, 0) * 5deg)) !important;
+  filter: contrast(1.2) brightness(1.3) blur(calc(var(--item-index, 0) * 1px)) drop-shadow(0 15px 25px rgba(0, 150, 255, 0.5)) !important;
+  z-index: calc(150 - var(--item-index, 0)) !important;
+  opacity: calc(1 - (var(--item-index, 0) * 0.15)) !important;
+  border-bottom: 3px solid rgba(0, 150, 255, 0.6) !important;
+  border-radius: 0 0 8px 8px !important;
+}
